# SwaggerOpenAPI_TaskTracker.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task Tracker API",
    "description": "Task Tracker API by Milos Maljenovic",
    "contact": {
      "name": "Milos Maljenovic",
      "email": "milosmaljen@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:55690/"
    }
  ],
  "paths": {
...
    "/api/tasks/project/{id}": {
      "get": {
        "description": "Get all Tasks for specified Project(Id)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success Response with Path Param",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tasksDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. Id is smaller than 0"
          },
          "404": {
            "description": "Not found. This project has no tasks attached."
          }
        }
      }
    },
...
    "/api/projects/{start}/{end}": {
      "get": {
        "description": "Get Projects who are in provided date range",
        "parameters": [
          {
            "name": "start",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok. Returns array of projects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/projectsDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. In case date are not provided or start date is younger than end date."
          },
          "404": {
            "description": "Not found. There is no projects in provided date range."
          }
        }
      }
    },
    "/api/projects/with/{priority}": {
      "get": {
        "description": "Get Projects who have provided priority",
        "parameters": [
          {
            "name": "priority",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok. Returns array of projects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/projectsDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. In case priority is smaller than 1 or greater than 3."
          },
          "404": {
            "description": "Not found. There is no projects with provided priority."
          }
        }
      }
    },
    "/api/projects/by/{status}": {
      "get": {
        "description": "Get Projects who have provided status",
        "parameters": [
          {
            "name": "status",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok. Returns array of projects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/projectsDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. In case status is smaller than 0 or greater than 2."
          },
          "404": {
            "description": "Not found. There is no projects with provided status."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "tasksDTO": {
        "type": "array",
        "example": [
          {
            "Id": 1,
            "Name": "Main task",
            "Status": 2,
            "Description": "Do main task",
            "Priority": 1,
            "ProjectId": 2,
            "ProjectName": "Project 2"
          },
          {
            "Id": 2,
            "Name": "Side task",
            "Status": 1,
            "Description": "Do side task",
            "Priority": 2,
            "ProjectId": 2,
            "ProjectName": "Project 2"
          }
        ],
        "items": {
          "$ref": "#/components/schemas/tasksDTO_inner"
        }
      },
...
      "projectsDTO": {
        "type": "array",
        "example": [
          {
            "Id": 1,
            "Name": "Project 1",
            "Status": 0,
            "Code": "Code001",
            "Priority": 1,
            "StartDate": "2019-02-22",
            "CompletionDate": "2019-06-22"
          },
          {
            "Id": 2,
            "Name": "Project 4",
            "Status": 1,
            "Code": "Code002",
            "Priority": 3,
            "StartDate": "2019-07-29",
            "CompletionDate": "2019-09-29"
          }
        ],
        "items": {
          "$ref": "#/components/schemas/projectsDTO_inner"
        }
      },
      "tasksDTO_inner": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer"
          },
          "Name": {
            "type": "string"
          },
          "Status": {
            "type": "integer"
          },
          "Description": {
            "type": "string"
          },
          "Priority": {
            "type": "integer"
          },
          "ProjectId": {
            "type": "integer"
          },
          "ProjectName": {
            "type": "string"
          }
        }
      },
      "projectsDTO_inner": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer"
          },
          "Name": {
            "type": "string"
          },
          "Status": {
            "type": "integer"
          },
          "Code": {
            "type": "string"
          },
          "Priority": {
            "type": "integer"
          },
          "StartDate": {
            "type": "string"
          },
          "CompletionDate": {
            "type": "string"
          }
        }
      }
    }
  }
}